Move styled-component to its own file
