+++ src/client/src/ui/spotTile/components/PriceButton/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './PriceButton'
